Refactor OperationsEvent creation
Among other, we now resolve the timestamp definition early and do most
of the validation up front.  Additionally, the entities are now built
as a single entity that JPA stores in one go.  Prior to this, the code
would create the event first and then linking the timestamp definition
as a separate step (which was getting awkward because the "link" now
stores data rather than being an SQL bridge between two tables).

Signed-off-by: Niels Thykier <nit@globeteam.com>